requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pymongo==4.6.0
motor==3.3.2
pydantic==2.5.0
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==4.1.3
python-multipart==0.0.6
python-dotenv==1.0.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
email-validator==2.3.0
openai==1.10.0
Pillow==10.2.0
langchain
langchain-openai==0.0.5
langchain-community==0.0.16
imagekitio